+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Use case-insensitive matching for MSBuild properties and macros to improve robustness and correctness.
+
 ### Fixed
 
 - Fix option --include-headers not including header files processed with Qt MOC.