+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Add support for MFC projects.
 - Add support for Win32 .rc resource files.
 - Support configuration-specific values for MSBuild property `TargetName`.