+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.4.0] - 2026-03-01
+
 ### Added
 
 - Add support for MFC projects.
@@ -80,7 +82,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 First stable release.
 
-[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/chausner/vcxproj2cmake/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/chausner/vcxproj2cmake/compare/v1.0.0...v1.1.0